refactor: windows: clean up and remove redundent code
